+++ tests/base.py
@@ -1,4 +1,6 @@
-from typing import Dict
+import time
+
+from typing import Dict, List, Optional
 from urllib.parse import parse_qsl, urlencode, urlparse, urlunparse
 
 from flask_testing import TestCase
@@ -135,6 +137,7 @@ class WsTestClient(GatovidTestClient):
     """
 
     clients = []
+    matchmaking_delay = 0.0
 
     def create_app(self):
         self.app = super().create_app()
@@ -163,3 +166,43 @@ class WsTestClient(GatovidTestClient):
         # Lo guardamos para poder "limpiarlo" más tarde
         self.clients.append(client)
         return client
+
+    def parse_json_args(self, args):
+        return dict((key, arg[key]) for arg in args for key in arg)
+
+    def get_msg_in_received(
+        self, received: List, msg_type: str, json: bool = False
+    ) -> (Optional[Dict], Optional[List[Dict]]):
+        """
+        Devuelve la primera aparición de un mensaje de tipo `msg_type` en
+        `received`.
+        """
+        raw = next(iter(filter(lambda msg: msg["name"] == msg_type, received)), None)
+        if raw is None:
+            return None, None
+
+        args = raw["args"]
+
+        if raw and raw.get("args") and json:
+            args = self.parse_json_args(raw["args"])
+
+        return raw, args
+
+    def set_matchmaking_time(self, delay: float):
+        """
+        Para los tests se parchea el tiempo de espera para el inicio de la
+        partida, evitando que se tenga que esperar a que acabe.
+        """
+
+        import gatovid.api.game.match
+
+        gatovid.api.game.match.TIME_UNTIL_START = delay
+        self.matchmaking_delay = delay
+
+    def wait_matchmaking_time(self):
+        """
+        Espera el tiempo de inicio de una partida, con un pequeño margen para el
+        procesamiento en el backend.
+        """
+
+        time.sleep(self.matchmaking_delay * 1.2)
